requirements.txt
anthropic
fastapi
uvicorn
Pillow
requests
serpapi
python-dotenv
htmx 